// src/common/constants/error-types.ts
export enum ErrorTypes {
  INTERNAL = 'INTERNAL',
  INVALID_OPERATION = 'INVALID_OPERATION',
  INVALID_PARAMETER = 'INVALID_PARAMETER',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  CONFLICT = 'CONFLICT',
  BAD_REQUEST = 'BAD_REQUEST',
  NOT_FOUND = 'NOT_FOUND',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
  NOT_IMPLEMENTED = 'NOT_IMPLEMENTED',
  SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
  GATEWAY_TIMEOUT = 'GATEWAY_TIMEOUT',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
}